+++ src/main/java/com/fundly/chat/entity/SelbuyMsgEntity.java
@@ -5,14 +5,13 @@ import com.persistence.dto.SelBuyMsgDto;
 import java.util.Date;
 
 public class SelbuyMsgEntity {
-    private Integer room_num;
+    private Long room_num;
     private String buy_id;
     private String pj_id;
     private Date dba_reg_dtm;
     private String dba_reg_id;
     private Date dba_mod_dtm;
     private String dba_mod_id;
-
     public SelBuyMsgDto toSelBuyMsgDto() {
         return SelBuyMsgDto.builder().pj_id(pj_id).buy_id(buy_id).room_num(room_num).build();
     }
